+++ packages/dropdown/src/react/__specs__/storyshots.spec.ts
@@ -3,11 +3,6 @@ import initStoryshots, {
   snapshotWithOptions
 } from '@storybook/addon-storyshots'
 
-jest.mock('@pluralsight/ps-design-system-util', () => ({
-  ...jest.requireActual('@pluralsight/ps-design-system-util'),
-  uniqueId: jest.fn().mockImplementation(prefix => prefix + 'unique-id')
-}))
-
 const createNodeMock = () => document.createElement('div')
 
 initStoryshots({