+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.0.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@8.0.6...@pluralsight/ps-design-system-drawer@8.0.7) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-drawer
+
+
+
+
+
 ## [8.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-drawer@8.0.5...@pluralsight/ps-design-system-drawer@8.0.6) (2020-12-11)