+++ packages/radio/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-radio@7.2.17...@pluralsight/ps-design-system-radio@7.2.18) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-radio
+
+
+
+
+
 ## [7.2.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-radio@7.2.16...@pluralsight/ps-design-system-radio@7.2.17) (2020-12-11)